refactor: use distro for dotnet-community env injection (#3142)
## Description

This is towards deprecating the device per otel-sdk in odiglet. it's
done for almost all languages, and only 2 left are `dotnet-community`
and `dotnet-legacy`. this PR is to migrate the `dotnet-community`.

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

~~- [ ] Added Unit Tests~~
- [x] Updated e2e Tests
- [x] Manual Testing
~~- [ ] Manual Load Test~~

## Kubernetes Checklist

Remove one more usage of the device
(`instrumentation.odigos.io/{{param.LIBC_TYPE}}dotnet-native-community`)
in favor of injection static env vars via the distro manifest.

## User Facing Changes

Internal refactor