refactor: use the nodejs-community starter from agent dir (#3153)
This is a small refactor.

The `nodejs-community` script is currently hosted in odigos repo which
make it more complicated to share with odigos enterprise. I also think
it's more appropriate to be hosted in the nodejs agent repo so to that
all nodejs related code is at the same place