CORE-64 Separte device plugin to another container (#3152)
## Description

The primary goal of this PR is to improve the resilience of the odiglet
pod by isolating the deviceplugin into a dedicated container. With this
change, even if the main container crashes or becomes unavailable, the
deviceplugin container continues to run, serve device registrations, and
enable scheduling of pods that the instrumentor has assigned devices to.

<img width="579" alt="image"
src="https://github.com/user-attachments/assets/4db7269c-d15a-40b2-9d70-9aea4f0a1d04"
/>


## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [ ] Added Unit Tests
- [x] Updated e2e Tests
- [ ] Manual Testing
- [ ] Manual Load Test

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [ ] Changes how Odigos interacts with Kubernetes
- [ ] Introduces additional calls to the API Server (potential
performance impact)
- [ ] New Query/feature supported in all the k8s versions supported by
Odigos
- [ ] Modifies Odigos manifests (addressed in both CLI and Helm)
- [ ] Changes RBAC permissions

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

- [ ] Users need to take action before upgrading
- [ ] Automatic migration will modify existing objects (backward
compatible)
- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users need
to be aware of
- [ ] Documentation updated accordingly