Odigos config hook fix (#3068)
## Description
This PR renames odigos-config to odigos-configuration globally.

The main reason for this change is that the old `odigos-config` had a
legacy `helm.sh/hook` annotation in its Helm template, and that caused a
few annoying issues:
- Helm didn’t manage that ConfigMap like the rest of the resources, so
we had to manually delete it during uninstall.
- There’s also been a bug between Argo CD and Helm that seems related to
this annotation. I wasn’t able to reproduce it consistently, but this
rename should help avoid it altogether.

And beside of that, a small nit: the `odigos.io/system-object=true`
label was added to the `odigos-gateway` ConfigMap cli manifest to keep
things consistent with the rest of the resources.

## Migration Process

#### CLI (odigos update command)
During the update process, the CLI reads the existing odigos-config
ConfigMap and preserves its content.
It then creates a new odigos-configuration ConfigMap with the same data.
The old ConfigMap is then deleted since it becomes resource with old
value in `odigos.io/config` label (see `DeleteOldOdigosSystemObjects`
func for more detailes).

#### Helm
On upgrade:
A new odigos-configuration ConfigMap is created automatically.
A post-upgrade Helm hook Job deletes the legacy odigos-config if it
still exists.

On uninstall:
A pre-delete Job ensures the legacy odigos-config is deleted before
uninstall completes.

### Short Summary of the Changes
1. Global renaming of odigos-config to odigos-configuration in both CLI
and Helm code.
2. nit: added the odigos.io/managed: "true" label to the odigos-gateway
ConfigMap cli resource.

### How Has This Been Tested?
Upgraded between Odigos versions and verified:
- Existing config is preserved.
- The instrumentation process keeps running smoothly.
- Ran odigos config and odigos diagnose locally to make sure things
still work as expected.

Tested Helm-based upgrade and uninstall in a cluster:
Confirmed that the legacy ConfigMap is properly deleted by the new Jobs.

---------

Co-authored-by: Amir Blum <amirgiraffe@gmail.com>