ci: ignore mint.json from the git status check (#3176)
## Description

The destination docs generation script applies a formatting to the
`mint.json` file.

Sometimes when we manually add doc pages (for example PR #3167), the CI
fails due to formatting change (see job
https://github.com/odigos-io/odigos/actions/runs/16250264278/job/45878946842?pr=3167),
this change is not relevant to the actual generation of docs, and should
not be a reason to fail the CI.

This PR makes sure we ignore the file in the git status check.