CORE-83 Not set label if device enabled (#3165)
## Description

This PR introduces a small change to prevent setting node affinity on
pods when the mountMethod is set to device.

The reason for this change is that the device-based mount method already
ensures that pods are scheduled only on nodes where the Odiglet is
installed. In other words, the presence of the device acts as a natural
scheduling constraint, eliminating the need for explicit node affinity.

In contrast, when using the k8s-host-path mount method, node affinity is
still required — it's the mechanism that prevents pods from being
scheduled on nodes without the Odiglet.

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [ ] Added Unit Tests
- [ ] Updated e2e Tests
- [ ] Manual Testing
- [ ] Manual Load Test

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [ ] Changes how Odigos interacts with Kubernetes
- [ ] Introduces additional calls to the API Server (potential
performance impact)
- [ ] New Query/feature supported in all the k8s versions supported by
Odigos
- [ ] Modifies Odigos manifests (addressed in both CLI and Helm)
- [ ] Changes RBAC permissions

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

- [ ] Users need to take action before upgrading
- [ ] Automatic migration will modify existing objects (backward
compatible)
- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users need
to be aware of
- [ ] Documentation updated accordingly