fix(scheduler): use force to apply odigos config if conflict (#3184)
#3068  changed `odigos-config` to `odigos-configuration`.

It was also changed in the `FieldManager` to a different value. since
the value is cahnged, we now get this error in scheduler:

```
Apply failed with 1 conflict: conflict with \"scheduler-odigosconfig\": .metadata.labels.odigos.io/profiles-hash"
```

This PR fixes the issue by using `Force: true`.

I reproduced it with a profile that has instrumentation rules and
actions, saw this error, then deployed the fix and observed it resolved