feat: calculate env injection method in reconcile  (#3135)
## Description

This is another step in the journey for deprecating otel-sdk and
env-overwite.

Today, pods webhook applies some logic deciding which env injection to
use, based on:
- odigos configuration (`agentEnvVarsInjectionMethod` which can have 3
values)
- runtime detection (`securedEcecution` and `LD_PRELOAD` in EnvVars)
- `otelSdk` as this logic will only run for languages found in this
static map in code.

The initiative is to move this calculation to the reconcile phase, and
only storing single enum on each container agent configuration to
instruct it what to use in injection:
- less complexity in webhook which is sensitive component that might get
run frequently
- better visibility in ui and instrumentation config to what odigos is
doing
- less dependencies in the webhook - consolidate all it's inputs into a
single place which gets updated atomically in reconcile which is less
room for race conditions and conflicts.
- one more place where we don't rely on otelSdk and migrate to use
distro manifests. this will help us deprecate it and cleanup our code a
bit.


This PR is the first step - write the env injection method into the
containers config so we can later use it in webhook

## How Has This Been Tested?

- [x] Manual Testing

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [x] Changes how Odigos interacts with Kubernetes
writes one more enum into each agent config in "InstrumentationConfig"
object

~~- [ ] Introduces additional calls to the API Server (potential
performance impact)~~
~~- [ ] New Query/feature supported in all the k8s versions supported by
Odigos~~
- [x] Modifies Odigos manifests (addressed in both CLI and Helm)
added to CRD

~~- [ ] Changes RBAC permissions~~

## User Facing Changes

Internal refactor.

---------

Co-authored-by: Ron Federman <73110295+RonFed@users.noreply.github.com>