chore(odiglet): simplify device plugin container in odiglet (#3199)
## Description

Odigos has 2 modes of mounting the `/var/odigos` directory into
instrumented containers: "k8s-host-path" and "virtual-device".

The "virtual device" is the default one, and used by most of odigos
users. Instrumentor webhook will add a "resource" request on
instrumented containers with the name of the device, and then k8s
scheduler will consider only nodes with this device as relevant for
instrumentation.

This mechanism is important for the stability of the cluster as it
affects k8s scheduler.

This PR simplifies the "device-plugin" container, removing potential
failure points from it that are no longer used or can be refactored,,
and reducing our failure surface for this critical component.

- Remove unused devices. enterprise still uses the dotnet legacy device.
after it is remove we can clean this up even further.
- remove the query to get `enabled signals` from the `collectors group`.
it was used by java virtual device which has migrated to generic many
months ago and is being removed in this PR.
- remove `collectors group` RBAC permissions for odiglet as they are no
longer needed.
- fallback to device amount of `1024` instead of setting it to
allocatable pods. the device is used per container, not per pod, and the
query is another fail point. we don't have an actual limit so using a
static number is good enough.

## How Has This Been Tested?

Run it manually and made sure for the changes in the manifest. saw apps
starting and traces.
this is a refactor and is tested in e2e.

## Kubernetes Checklist

- [x] Changes how Odigos interacts with Kubernetes
removed many devices which will not be registered to the node anymore.

~~ [ ] Introduces additional calls to the API Server (potential
performance impact)~~
~~ [ ] New Query/feature supported in all the k8s versions supported by
Odigos~~
~~ [] Modifies Odigos manifests (addressed in both CLI and Helm)
- [x] Changes RBAC permissions
removed `CollectrosGroup` from odiglet as it is not used anymore

## User Facing Changes

Internal refactor. removing unused things. no impact on user