ci: create pre-release action for release PR (#3218)
Add a new action to create release PR from CI.

It will calculate the next stable release, then the next stable
release-candidate version, and will create (if required) a release
branch for the stable version, and a PR to create release candidate from
this branch.

This is first iteration, we still need to tag the repo when the release
PR is merged.