ci: add full build for release candidate prs (#3221)
Add a workflow to build all the services, with both rhel and regular
dockerfiles, and for both arm64 and amd64.

We are currently doing this on every PR to main, and I want to do it for
the release-candidate PRs as well.

Later on, we can consider to run the full builds only on release PRs,
and avoid them on regular merges to main, as they take a lot of time to
run, cost a lot of money, and give very little value.