fix: daemonsets fails to create instrumentation config (#3227)
## Description

Odigos fails to create instrumentation config CR for source of type
Daemonset.
This is due to a bug introduced in #3163 which limits the daemonsets we
pull into the cache in instrumentor to only "odiglet".

This PR revert this bug to make sure all daemonsets are found in cache
and can be reconciled correctly

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

Manually

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

Revert change, so now we pull all daemonsets into the instrumentor
cache.
The RBAC permissions are ok (cluster role for all required operations)

## User Facing Changes

Bug fix - for any user that is instrumenting daemonsets