autoscaler: legacy actions in K8sAttribute processor ownerreferences (#3272)
## Description

Fixing a bug introduced in https://github.com/odigos-io/odigos/pull/2951

The way K8sAttributes actions work is that all Actions CRDs
(legacy+generic) are combined into a single Processor. This Processor
has the individual Actions as OwnerReferences.

In the current code, we convert all the legacy Actions to generic
Actions to simplify the code for generating the Processor config.
However, this messed with setting the OwnerReferences for the legacy
Actions as they were being converted to a different API type and their
UID wasn't preserved. Leading to:

```
Action Transformed To Processor Type
Failed to transform the action to a unified processor: Processor.odigos.io "odigos-k8sattributes" is invalid: metadata.ownerReferences.uid: Invalid value: "": uid must not be empty
```

This updates the OwnerReference part to add all objects, legacy+generic,
to the new Processor

Thank you @tamirdavid1 and @AvihuHenya for reporting and helping with
some debug

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [ ] Added Unit Tests
- [ ] Updated e2e Tests
- [x] Manual Testing
- [ ] Manual Load Test

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [ ] Changes how Odigos interacts with Kubernetes
- [ ] Introduces additional calls to the API Server (potential
performance impact)
- [ ] New Query/feature supported in all the k8s versions supported by
Odigos
- [ ] Modifies Odigos manifests (addressed in both CLI and Helm)
- [ ] Changes RBAC permissions

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

- [ ] Users need to take action before upgrading
- [ ] Automatic migration will modify existing objects (backward
compatible)
- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users need
to be aware of
- [ ] Documentation updated accordingly