fix: build rsync in multi-arch stage (#3278)
When building the `rsync` binary we used the 
```
FROM --platform=$BUILDPLATFORM registry.odigos.io/odiglet-base:v1.8 AS builder
```

Since our release flow runs on amd64 this means that this step runs on
amd64.
Hence when calling `make` it compiles it to `amd64` even when the target
architecture is arm64.

This is fixed by separating the rsync build step, and not using an image
pinned to the build platform - this way this step will run with both
amd64 and arm64 images based on the target architecture.