fix: do not apply HTTPS_PROXY on k8s apiserver and own metrics (#3305)
## Description

Following https://github.com/odigos-io/odigos/pull/3296, it was found
that the requests to the api server where also capture and using the
https proxy, which made it fail.
This PR makes it so that connecting to api server never uses a proxy,
fixing the above issue.

While testing, it was found that gRPC connections, even when not
encrypted, are using the HTTP CONNECT flow which is also using the
HTTPS_PROXY definitions and breaks the own telemetry reporting to the ui
pod. this is fixed by adding it to the `NO_PROXY` list and black-list it
so proxy is disabled for it.

## How Has This Been Tested?

Setup an https proxy, set this option, and made sure the requests are
passing via the proxy and gateway logs are ok

## Kubernetes Checklist


## User Facing Changes

bug fix that does not impact user