CORE-127: Add OverwriteExistingValues option to AddClusterInfo action (#3306)
## Description

Currently, we can only add new cluster info attributes since we hard
code `insert` as the action in the resource processor. Allowing users to
toggle to `upsert` gives the option to overwrite existing attributes.

A toggle was chosen for better UX, as the details of "insert" vs
"upsert" may be overly technical (and users do have the option of
manually creating a Processor CRD anyway).

Similarly, this option applies to the entire Action instead of
individual attributes for simpler UX.

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [ ] Added Unit Tests
- [ ] Updated e2e Tests
- [ ] Manual Testing
- [ ] Manual Load Test

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [ ] Changes how Odigos interacts with Kubernetes
- [ ] Introduces additional calls to the API Server (potential
performance impact)
- [ ] New Query/feature supported in all the k8s versions supported by
Odigos
- [ ] Modifies Odigos manifests (addressed in both CLI and Helm)
- [ ] Changes RBAC permissions

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

- [ ] Users need to take action before upgrading
- [ ] Automatic migration will modify existing objects (backward
compatible)
- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users need
to be aware of
- [ ] Documentation updated accordingly