fix: set controller reference on collector group (#3311)
## Description

Collectors group CR currently has no owner reference, even thought it is
controlled by a controller (scheduler).

This causes an issue that if argoCR is redeploying odigos in a "clean"
(prune) manner, then helm will delete all the objects it controls and
recreate them, but the collectors group and the collectors themselves
are left untouched.
Setting the controller reference guarantees that if for any reason the
scheduler is deleted, so are the resources it owns and control. When
scheduler is re-created, it will reconcile the objects which will
re-create the collectors groups and re-deploy the collectors as
required.

## How Has This Been Tested?

- [x] Manual Testing

## Kubernetes Checklist

- [x] Changes how Odigos interacts with Kubernetes - adds a new owner
reference.
- [x] Changes RBAC permissions - add deployment RBAC to scheduler in
both helm and cli

## User Facing Changes

Bug fix for very specific deployment sequence