feat: add config option to check device plugin health befoe injection (#3209)
## Description

This adds a new opt-in option called "checkDeviceHealthBeforeInjection".
When this option is set, and the mount method is set to
"k8s-virtual-device", instrumentor webhook will do the following checks
before injecting odigos into the pod manifest of an instrumented pod:

- odiglet ds is found (e.g. not deleted)
- all odiglet pods, "deviceplugin" container are not in
"CrashLoopBackoff" state.

If at least one odiglet device plugin container is in crash loop, the
instrumentation is skipped.

## How Has This Been Tested?

Made the device plugin container crash in code (added panic), then
uploaded it to the cluster, set this config option, and verified that it
does not inject instrumentation into the pod.

Also verified:
- when config option changes, it affects the pods webhook correctly
- when odiglet device plugin container status changes (is crash loop and
now not, or vice versa), then new pods are injected instrumentation or
not based on this value.

## Kubernetes Checklist

This change requires for instrumentor to pull all pods into the cache
(not just instrumented pods).

- rollback logic needs all instrumented pods (with odigos label present)
- pods webhook needs the odiglet pods (not instrumented, but with the
`app.kubernetes.io/name: odiglet` label.

k8s does not allow to set selectors on "OR" conditions, or to mix
"Exists" with "exact value" for labels.

This will increase instrumentor memory consumption (specially during new
installations before any pod is instrumented).

instrumentor transform pods and strip out most of the data (only keeps
container status) so it's not expected to be an issue, but it is a
change to monitor.

Also, pods webhook will need to work a bit harder to list and check all
odiglet pods everytime it is invoked

## User Facing Changes

New feature for enhanced stability in the cost of sometimes missing
instrumentation.