Avoid adding deployments/replicaset metadata in logs pipeline by default (#3336)
Using the `k8sattributes` processor - collecting deployment metadata can
be memory expensive in big clusters. This is because the processor will
pull into a cache all the replicasets from the cluster. As opposed to
pod metadata this can't be filtered to only pull the node objects.

For logs, disable this collection by default. For traces we have these
attributes from each instrumentation so no need to add them in the
collector.

If a user wants to keep those attributes for logs, it still can be done
by adding the `k8sattributes` action.