[RUN-112] Reduce hostPath mounts, permissions and hostPID for odiglet and data collection (#3331)
Adding opt in options to allow odiglet to run with less privileges and
less mounts.
* Reducing host mounts. When opt in - we only allow the `init-conatiner`
mount method. This is because `device` and `hostPath` mount methods
assumes `/var/odigos` is present on the host and can be shared with
relevant containers. Two exceptions which are made are `/proc` and
`/sys/kernel/debug`. The first can be mounted into the odiglet and by
that remove `hostPID: true`. The second is required for attaching
tracepoints. Both of them are used by most popular and mainstream eBPF
tools. Additional host mount that odiglet does is `/run` - this is for
accessing CRI sockets - this is not a mandatory feature and we can deal
without it when opt in. Can be further fine tuned in the future - for
example by only mounting `/run/containerd`.
* hostPID is removed by mounting the `/proc` folder from the host.
Different odiglet components are updated to get the proc filesystem
location as a configuration. This includes the runtime-detector and the
different language inspectors.
* privileged is replace with the required capabilities. BPF - load and
manage eBPF programs and maps. SYS_PTRACE - perform ptrace in some of
the instrumentation libraries - and also required for access some of the
proc files. PERFMON - required to load tracepoints in the
runtime-detector. DAC_READ_SEARCH - required for some of the proc files
like `/mpas` and `/environ`.

changes include:
* odiglet helm template to take into account the 3 new values. All the
new values are opt-in - and the current behavior should not be changes.
* bump `runtime-detector` to v0.0.18 - this version includes an
additional option to configure the proc fs location.
* update instrumentor helm template to fix a bug in the `odigos-agents`
images being used - add a condition to take the enterprise image when
needed.