perf: optimize source fetching (#3384)
We added pagination to fetch sources in the past because even a few
sources would take a while to fetch - that was when we would get
instances in the same request.

We've since separated the requests, and Sources are much faster to
query, even to the point where pagination over-complicates things.

So now, we have to remove pagination :)

---

This PR also bumps the UI Kit with related performance improvements:
https://github.com/odigos-io/ui-kit/releases/tag/ui-kit-v0.0.98

---------

Co-authored-by: Alon Braymok <138359965+alonkeyval@users.noreply.github.com>