CORE-169: Use NODE_IP env var for kubeletstats receiver (#3387)
## Description
This PR updates the `kubeletstats` receiver configuration to use the new
`NODE_IP` environment variable, in order to support environments where
the NODE_NAME value is not resolvable in cluster DNS.

<!-- Short summary of the changes -->
* Added a new NODE_IP environment variable in `data-collection`.
* Updated the auto-scaler nodeCollector controller to create the
`kubeletstats` receiver based on the NODE_IP value.

## How Has This Been Tested?
Configured a Prometheus destination in the cluster and verified that
kubelet metrics were successfully fetched.

---------

Co-authored-by: Tamir David <tamir@odigos.io>