fix(collector): increase the livness probe to 5 sec (#3392)
Fixes: CORE-188


Our gateway collector, under load, will have it's CPU and memory under
pressure. If the CPU is the limiting factor, then the endpoint for the
health_check might take time to response, and can timeout the kubelet
caller, and fail the check.

After 2 failed check the collector will be restarted.

This PR increase the value to 5 seconds (which is still an arbitrary
number, but less likely to fail).