fix: check memory limit using runtime gc controller (#3359)
Fixes: CORE-192

This PR targets the collector crashing due to OOM under high load or
spikes.

Previous to this PR, the memory limit was enforced using the
OpenTelemetry Collector memory limiter extension to periodically check
if the limit is reached and respose by refusing data with RESOURCE_
EXHUSTED in this case.

OpenTelemetry uses the `runtime.ReadMemStats` function which gives
consistent memory stats, but it also stops the world and is expensive to
call, thus cannot work for high-frequency, per request checks.

It is replaced by similar code that:
- `go:linkname` directly into the gc controller to fetch the relevant
values fast. The gc controller uses `atomic` to sync the access to these
values.
- hooks into the same computation the go GC is making, to make sure we
end up with the same result.