refactor: set cri error on runtime detection only when there is actual error (#3437)
Fixes: CORE-206

This PR is for a change in semantics for the `CriErrorMessage` field of
`RuntimeDetailsByContainer`.

Until now, we would always set it if a CRI error is detected. 
There is a branch, however - we know that env vars from container
runtime will be found in `/proc`, so we check and if we don't find them
there, we know that `envFromContainerRuntime` is empty even if there is
an error.

Currently we set `CriErrorMessage` to the error message, and use the
`runtimeUpdateState` to differentiate between the other case. This is
what the webhook checks for injection.
Since the enum was meant to be temporary, and there is an ongoing effort
to refactor the webhook code to make it use distros and retire the
`EnvOverwriter` and `OtelSdk`, I want to introduce this change towards
removing it and encode it's actual decision into the error message
instead.

The `criErrorMessage` is currently not used for anything other than
print message to log, thus this change is safe. [Webhook
code](https://github.com/odigos-io/odigos/blob/15b77f64310b5202616d4842f78b96f7d65d706f/instrumentor/internal/webhook_env_injector/webhook_env_injector.go#L239)
only uses the `RuntimeUpdateState` which is unchanged.

It's worth noting that we have never actually seen a criError in any
deployment since it was introduced 8 months ago, so this change should
be safe in that regard as well