chore: consume nodejs-community distro aratifact as versioned dependency (#3438)
Fixes: CORE-200

Consume the `nodejs-community` distribution artifacts (the agent
implementation) as a docker image with version number:

Pros:

- No new code is introduced into the build by pushing to upstream `main`
branch (as we do now)
- We can publish from release-branch and be deterministic about the
version that is being used regardless of agent repo pushes.
- Faster builds - no need to build the agent from scratch in odiglet and
init-container agents image. simply copy the existing artifacts (6.5 MB)
from a container where it's already ready to be used.

Cons: 

- The convenient `build-odiglet-with-agents` does not work with this new
setup
- we need to remember to release the agents in the many source repos,
and then bump the version here (in 5 places). Can be automated in the
future, but more chore and option to forget and make mistakes.
- Agent version number is now independent of odigos and may be a bit
harder to correlate the version in trouble-shooting